Proj_1/requirements.txt
Django==1.9.13
django-crispy-forms==1.7.2
django-filter==2.2.0
django-markdown-deux==1.0.5
django-pagedown==1.0.6
markdown2==2.3.8
Pillow==6.2.1
pytz==2019.3
sqlparse==0.3.0
